aplicar policies tanto en la vista como en el controlador

+++ app/Http/Controllers/CommentController.php
@@ -36,7 +36,7 @@ class CommentController extends Controller
 
     public function edit(Post $post, Comment $comment)
     {
-        //$this->authorize('update', $comment);
+        $this->authorize('update', [auth()->user(), $comment]);
         $parent_comment = $comment->parent;
 
         return view('comment.edit', compact('comment','parent_comment', 'post'));
@@ -44,7 +44,7 @@ class CommentController extends Controller
 
     public function update(UpdateCommentRequest $request, $post, Comment $comment)
     {
-        //$this->authorize('update', $comment);
+        $this->authorize('update', [auth()->user(), $comment]);
         $comment->update($request->validated());
 
         return to_route('post.show', $post)
@@ -53,7 +53,7 @@ class CommentController extends Controller
 
     public function destroy(Post $post, Comment $comment)
     {
-        //$this->authorize('delete', $comment);
+        $this->authorize('delete', [auth()->user(), $comment]);
 
         $comment->delete();
